Changed mongoDB Authentication
Now forced a security method to be able to log into the database, from only the server so that no one else can be able to authenticate into it. This allows us to move the server so that it doesn't have to be only localhost if we want to use a remote server.