Created Testable Functions, but REMOVED a lot of features.
A lot of functions were removed in order to simplify and understand how to design testable functions. BaseInformation can now be unit testable, if you've provided a mock Collection and mock methods.
We need to readd the other features, but keep in mind of how to test the functions. This is the beginning of TDD.