Added Update button to profits page
New update button that will update when clicked on every 12 hours. We won't automate this process, because it will hit the API too many times in a day. Rather, we will just create an update function when users request after a SEARCH. Next up is to put this update link on the OBTAIN page.